incubator-todolist-ts-09/src/state/tasks-reducer.test.tsx
import { tasksReducer, RemoveTaskAC, AddTaskAC, ChangeTaskStatusAC, ChangeTaskTitleAC } from './tasks-reducer';
import { TasksStateType } from '../App';

test('correct task should be deleted from correct array', () => {
    const startState: TasksStateType = {
        "todolistId1": [
            { id: "1", title: "CSS", isDone: false },
            { id: "2", title: "JS", isDone: true },
            { id: "3", title: "React", isDone: false }
        ],
        "todolistId2": [
            { id: "1", title: "bread", isDone: false },
            { id: "2", title: "milk", isDone: true },
            { id: "3", title: "tea", isDone: false }
        ]
    };

    const action = RemoveTaskAC("2", "todolistId2");

    const endState = tasksReducer(startState, action)

    expect(endState["todolistId1"].length).toBe(3);
    expect(endState["todolistId2"].length).toBe(2);
    expect(endState["todolistId2"].every(t => t.id != "2")).toBeTruthy();
});

test('correct task should be added from correct array', () => {
    const startState: TasksStateType = {
        "todolistId1": [
            { id: "1", title: "CSS", isDone: false },
            { id: "2", title: "JS", isDone: true },
            { id: "3", title: "React", isDone: false }
        ],
        "todolistId2": [
            { id: "1", title: "bread", isDone: false },
            { id: "2", title: "milk", isDone: true },
            { id: "3", title: "tea", isDone: false }
        ]
    };

    const action = AddTaskAC("new", "todolistId2");

    const endState = tasksReducer(startState, action)

    expect(endState["todolistId1"].length).toBe(3);
    expect(endState["todolistId2"].length).toBe(4);
    expect(startState["todolistId2"].length).toBe(3);
    expect(endState["todolistId2"][0].title).toBe("new");
});

test('correct task should be change status', () => {
    const startState: TasksStateType = {
        "todolistId1": [
            { id: "1", title: "CSS", isDone: false },
            { id: "2", title: "JS", isDone: true },
            { id: "3", title: "React", isDone: false }
        ],
        "todolistId2": [
            { id: "1", title: "bread", isDone: false },
            { id: "2", title: "milk", isDone: true },
            { id: "3", title: "tea", isDone: false }
        ]
    };

    const action = ChangeTaskStatusAC("2", "todolistId2");

    const endState = tasksReducer(startState, action)

    expect(startState["todolistId2"][1].isDone).toBe(true);
    expect(endState["todolistId2"][1].isDone).toBe(false);
    expect(endState["todolistId2"].length).toBe(3);
    expect(startState["todolistId2"].length).toBe(3);
});

test('correct task should be change title', () => {
    const startState: TasksStateType = {
        "todolistId1": [
            { id: "1", title: "CSS", isDone: false },
            { id: "2", title: "JS", isDone: true },
            { id: "3", title: "React", isDone: false }
        ],
        "todolistId2": [
            { id: "1", title: "bread", isDone: false },
            { id: "2", title: "milk", isDone: true },
            { id: "3", title: "tea", isDone: false }
        ]
    };

    const action = ChangeTaskTitleAC("2", "todolistId2", "new");

    const endState = tasksReducer(startState, action)

    expect(startState["todolistId2"][1].title).toBe("milk");
    expect(startState["todolistId2"].length).toBe(3);
    expect(endState["todolistId2"][1].title).toBe("new");
    expect(endState["todolistId2"].length).toBe(3);
});

